python/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling>=1.27"]

[project]
authors = [{ name = "Yunho Cho", email = "ycho358@gatech.edu" }]
dependencies = [
  "numpy>=2.3,<3",
  "onnx>=1.19,<2",
  "opencv-python-headless>=4.12,<5",
  "pillow>=11.3,<12",
  "pyclipper>=1.3.0.post6,<2",
  "pyyaml>=6.0.2,<7",
  "requests>=2.32,<3",
  "shapely>=2.1,<3",
  "tqdm>=4.67,<5",
]
description = "Python package and Pixi workspace for PaddleOCR-TVM."
name = "paddleocr-tvm"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"

[project.scripts]
paddleocr-tvm = "paddleocr_tvm.cli:main"

[project.urls]
Repository = "https://github.com/yunho-c/PaddleOCR-TVM"

[dependency-groups]
dev = [
  "mypy>=1.18,<2",
  "pytest>=8.4,<9",
  "pytest-cov>=7,<8",
  "ruff>=0.13,<0.14",
]
parity = [
  "paddlepaddle>=3.1,<4",
  "paddlex>=3.4,<4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/paddleocr_tvm"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-arm64", "win-64"]

[tool.pixi.pypi-dependencies]
"paddleocr-tvm" = { path = ".", editable = true }

[tool.pixi.dependencies]
onnxruntime = ">=1.24,<2"

[tool.pixi.environments]
default = { features = [], solve-group = "default" }
dev = { features = ["dev"], solve-group = "default" }
parity = { features = ["dev", "parity"], solve-group = "default" }

[tool.pixi.tasks]
ocr-mobile = "paddleocr-tvm ocr-mobile"
prepare-mobile = "paddleocr-tvm prepare-mobile --artifacts-dir .artifacts --target llvm"
start = "python -m paddleocr_tvm --help"

[tool.pixi.feature.dev.tasks]
check = { depends-on = ["lint", "typecheck", "test"] }
format = "ruff format ."
lint = "ruff check ."
test = "pytest"
test-cov = "pytest --cov=paddleocr_tvm --cov-report=term-missing"
typecheck = "mypy src"

[tool.pixi.feature.parity.tasks]
parity-mobile = "paddleocr-tvm parity-mobile"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.mypy]
packages = ["paddleocr_tvm"]
python_version = "3.11"
strict = true
